WIP: using epoll for multiclient support
Actually multi client are supported until they start to write something.
As soon as a client starts sending a command, this client is fully
blocking until the connexion is closed.

This is a work-in-progress: the workflow is still the same as the
previous implementation except that epoll is already half-implemnted to
allows better implementation soon.