Major refactor introducing early namespaces support
Previously, a single index and single data structure was kept in memory
for handling every call done, using one global variable, since only a
single index and a single data was used globally

To make things more clean, get ride of the global variable was for sure,
needed. Now the index and the data structure are passed each time to the
function which needs them.

Now, with this change, we can support multiple index and multiple
dataset, which allows us to introduce namespace notion.

Previously, the main was building an index and data, then listening for
sockets commands, now everything is given to the namespace manager.

There is always at least one namespace (the default one, called
'default'), which replace the previous single-one namespace used. Like
this the logic doesn't change too much since we always have at least one
we can use as fallback. The code impact is only the replacement of the
global variable by local variables and arguments.

The main process invoke a namespace manager which will (for now) load
the default namespace, and keep per namespace, the index and data
structure. As soon as a client connects, it's attached to the default
namespace, and this namespace links to index and data structure.

Multiple client can use the same namespace, they will all points to the
same namespace structure, which points to the same index and data
structure, like this, we only have one copy of the structures whatever
the amount of clients.

For now, only the default namespace is loaded, any others namespace are
ignored and we can't create a new one, but the code is there, and the
new logic is here. This change should change nothing in a runtime point
of view now, except that a new sub-directory called 'default' will
appears on the index and data root directory.