Add security layer on direct-key mode when requesting key
Since the user can request a key which contains binary information about
the location of the data, the user could request any key with malformed
data or with arbitrary requested offset/fileid invalid or not
legitimate.

This commit fix this by adding an extra security check on direct-key
mode only. When requesting a key, some security points are checked:
 - Does the file id requested exists
 - Can we read data at the requested offset
 - Does the header at that offset correspond of the requested key

Only if all of theses points are validated, the data are fetched