Introduce data header flags to flag as deleted (implment #15)
Add a new field on data header: flags
Like index, this flag is used for now to mark a block as deleted.

Since we can use the data files to rebuild an index, this is useful in
that case.

More over, in direct-key mode, we don't rely on the index at all, this
is the only way to know a key was deleted.

The default behavior is always append, but in this case there is no
other was to handle correctly delete for direct-key mode without
overwriting the existing header, this is the only exception when a
closed data file can be re-opened in read-write and not in append mode.