Massive argument change (resp -> client, related to #11)
Sending 'client' for each command request, in place of the previous
'request'

Previously, each request got a link to the client, now the client
contains a pointer to the request (there can be only one request at a
time).

This makes everything more flexible for later.
Attaching one request per client (and not a client to a request) makes
it easier to make client fully non-blocking, the state of the request
will be kept per client.

This is related to #11 because this will make the fix easier, but it's
not yet fixed, this is only a major code refactoring to prepare the next
change.