Remove any blocking socket code (basicly finish #11)
There are no any blocking socket anymore.

When a client connects, the socket is set in non-blocking mode (like
before) and the status never change after that. This fixes the possible
lockdown issue by buggy clients or malformed clients.

The new implementation keeps one small buffer per client, and the
current state (status of the work), by doing this, each time something
is received on the socket, we parse this small amount of data, and keep
going for something else as soon as new data are available. One client
cannot block another one, except during the execution of the command of
course.