Always allocate single object in index
Always allocate single and reusable object, whatever mode is selected.

For future plans (like #31), mode could not be unique, let's allocate
everything we need (it's all small) and keep it a life-cycle long and
clean them when we're done.

This, btw, fix a non-free'd variable on STOP in (previously) direct
mode. But this was not a memory leak.