Preload indexfile in memory before populate it (fixes #34)
The index load time was affected by lot of syscall (reading each entries
on disk, one by one). This was speedup a lot with this easy fix.

Since index files are always small files (eg: ~20 MB for 256 MB datafile
of 1 KB per entries), we can load it in memory in one shot, then parse
this buffer. This is of course lot faster.

In addition, this simplify a lot the code since we don't need to
allocate an object and reading the length to know the amount of data to
allocate, this improve the loadtime and make code more easy.

Quick benchmark was made, for loading 7500154 entries (total of 14889.97
MB datasize), using sha256 keys, only the load time is computed:

- Debug mode, original code: 7.32s
- Release mode, original code: 6.14s

- Debug mode, new code: 3.81s
- Release mode, new code: 2.62s

Load time is a little bit more than twice as fast, and can be even more
with lot of small entries.

This closes #34