Improve SCAN and how key are managed on the index
Keeping offset on the index file in memory, for each entries.
This is needed to find the entry back on the index file, and flag the
entry as deleted when requested.

Since we use the index file to SCAN/RSCAN, we need to know if the entry
is deleted or not, appending new entry won't help in this case.

In addition, we keep the offset of the previous entry on the index, like
we did on the datafile, for the same goal