Add long-poll command notification support
Add a new 'WAIT' command, making the client connection to go in
background, waiting for an event. An event is anothing client executing
a specific command on the same namespace as the waiting client.

After each command, if a client in the same namespace was waiting for a
command executed by someone else, the waiting command receive '+OK'

Doing 'WAIT WAIT' will immediatly returns

Do wait on a command, you just need to 'WAIT' with command name as
argument, eg: 'WAIT PING'

This is related to #50