Major refactor and add support of duplication detection
This commit mostly contains code refactoring to reduce arguments length
and improve readability.

Long parameters and arguments when adding data into index and datafile
use intermediate struct, making arguments more verbose, clear and reduce
arguments to pass to parameters (sadly, this can lead to forget some
parameters and increase error not detected by compiler).

The index now contains the CRC32 field too, and is used to compare when
adding data, if the CRC match, we don't append the data, since it's
already there and nothing changed.

Maybe the date should be updated, this needs to be discussed.

This commit implement #26