Major change how deleted keys are managed in memory
Solve multiple bugs, change how keys are maintained in memory.

When deleting a key now, we remove the entry from memory, previously it
was only flagged and kept. This was wasting memory and causing bug when
ensuring does key was deleted or not when doing lookup.

This can cause side effects bugs.