Rewrite and improve protocol parser (support pipeline)
This is an important rewrite of the redis protocol parser from socket
and memory/buffer perspective.

This fix allows pipeline request (multiple request in a single packet)
and fix multiples issue when specific edge case occures (for example
when header of payload was at the end of a packet or in the middle of
two packet).

This should not break anything else and only provide better support of
redis protocol and edge cases, tests are adapted (next commit) and new
one was introduced to tests theses tricky cases. In addition some
external real-world usage was made to ensure it works.