Implement the SCAN command with bunch of keys response
The SCAN command now reply with multiple keys and not only a single one
(the only one next). The SCAN format follows the same behavior as before
except that second value of the array is again another array.

The implementation is made in a way which fetch as many keys as possible
in a fixed period of time (hardcoded to 1ms). It could be longer than
1ms if the last fetch is longer, but at least, no more fetch will be
done after 1ms, to avoid blocking others clients.

Here is an example of response:
> SCAN
1) "key-3"
2) 1) 1) "key-1"
      2) (integer) 16
      3) (integer) 1535359739
   2) 1) "key-3"
      2) (integer) 96
      3) (integer) 1535359814