Move index get to dedicated public api and file
Previously, each time an index entry was requested, the redis function
pointer based on the running mode was used, which was ugly and not
clean/portable for public api usage for later.

This is a first change to the code base to isolate this function pointer
related to the internal workflow of the index, and expose a public api
access to request a key entry from the index, whatever is the running
mode, it's not the caller which needs to adapt itself, but the index
handler.

Now, there is a 'index_get(index_root_t *, void *id, uint8_t idlength)'
function which take care of returning an 'index_entry_t *' based on the
running index, nothing more.