Complete rewrite and refactor of the index set handling
This is an important rewrite of the code handling how SET are done
(basicly all insertion/update inside index)

In prevision of next work, which try to separate redis and database
functions, to try to separate the work to make dedicated library to use
the backend, there was an important change to do, since the set command
logic between index, data and running mode was completely mixed on the
same bunch of code.

This new way make some isolation and allows 'redis set command' and
'internal usage of the index' to be isolated and more clean.

This introduce the 'index_set' files, which takes care to do the right
actions on the index, based on the running mode (of the index, not
global variable, which was a bad dependency).

By the way, the sequential code was fixed and update works as expected
and history still works.

This introduce new 'index_set' and 'index_set_memory' public function to
be used to update an entry on the index with a simple central function