Add staging KSCAN command (disabled in release) and generic list
Add KSCAN command in debug mode, which returns a list of keys matching
prefix provided as argument. This feature needs to iterate over the full
keyset, which is slow. Pagination like SCAN and time limit needs to be
added before adding this in release mode, right now this is a potential
easy way to slow down and freeze the server.

Btw, this include a list_t generic list support, to improve which will
possibly replace other list used on the code.