Improve how WAIT works and allows waiting on wildcard
Previously, WAIT was waiting for a specific command and returns +OK when
that command was triggered, this behavior changed

As before you can WAIT on any specific command but moreover you can now
waits on '*' command which is a wildcard and allows you to waits on any
commands.

In addition, when a watching command (or any if wildcard is used), the
command name is returned and not OK anymore.

Assume you 'WAIT PING', if somebody else do PING, you'll receive +PING
on the waiting side.

If you 'WAIT *' and somebody do 'INFO' for exemple, you'll receive +INFO