Introduce MIRROR command used for replication (staging)
A new MIRROR command made for special replication come in game. This
command basicly start streaming the full input to a client. This stream
contains more data than just what was received, it contains the socket
fd and the namespace attached to the client, because all of this are
state related to original client and we need it to replicate correctly
stuff.

As soon as you enter in MIRROR mode, you start receiving in streaming a
full copy with more context, this will be described later on
documentation, this is an early implementation, more stuff need to be
added to have a full replication working