Fix index inconsistancy when overwriting key in sequential mode
This fix an inconsistant change when overwriting a key in sequential
mode. When a key is overwritten, a new entry is added (flagged as
deleted) and the original one is updated to match with latest data and
history pointers.

The real problem is, the 'previous' field which point to the previous
offset in the index file (to reach the previous key for backward read)
was updated to match with the new key, which is completely wrong, the
original entry doesn't move, this broke the SCAN and RSCAN feature since
you could not move correctly sequentialy in the file anymore.

This change the content of the index file. A dirty fix will be applied
to future version to support this retro-compability with old indexfile.