Fix SCAN and RSCAN previous field issue (fixes #69)
This commit change internal workflow of SCAN and RSCAN feature and
improve a lot some features.

In previous commit, a bug was discovered and fixed, needing index
content change because of inconsistancy overwrite of 'previous' field
which was breaking linear reading of index files.

This commit add extra verbosity for debugging more easily serialization
and deserialization of keys when doing SCAN / RSCAN / KEYCUR.
This commit also introduce a way to keep track in which indexfile the
previous offset is located and this information is used on the reply of
SCAN / RSCAN / KEYCUR, this was a huge bug.

On previous version, the datafile and indexfiles were always inline for
enteries, any entries in a separate index file was always maching
content in the same dataid file (eg: entry in index-00004 was always
pointing to data in data-00004 file). Since we update the index file in
sequential mode to keep track of deleted keys and keeping
forward/backward linear read, this was not true anymore and was not
noticed. Now, a key inserted in the begining using index-00000 file and
updated later when multiple files exists could point to data in
data-00005 for example.

Code was stil assuming files id were inline, which result of breaking
reading (SCAN of KEYCUR was responding a key pointing to the wrong
indexfile because it was based on datafile, the indexfile was not stored
because not used).

This commit also introduce a backward compatibility with files created
with a buggy version of 0-db with wrong 'previous' field. Since this bug
only affect sequential mode and keys in sequential mode are hardcoded to
4 bytes, we don't really need to rely on the 'previous' field, we can
compute the previous key ourself, this is a dirty fix because this add
special case to take in account and not just use the generic way, but
backward compability is more important for production servers running
right now.