Add 'next_internal_id' in NSINFO response
In order to know what's the next internal id which will be generated on
the next data insertion, you'll find this information in NSINFO
response.

Value is in hexadecimal. If next id is '0xeb631000', next SET in
sequential mode will returns you '\xeb\x63\x10\x00'

Please note, this is an incremental id, but in little-endian.
Internal id is stored in 64 bits, but value returned by SET and
stored in sequential id are 32 bits.

This closes #71