Early mostly-working version of 'index-rebuild'
First partial working version of 'index-rebuild' using libzdb. This
version is a PoC of an index rebuild of a single data-file (for now)
without the history chain, but everything else is working.

There are so many less line of code by using the libzdb which takes care
of handling all critical stuff automatically.

This is one big step for maintenance tools.