Use internal index system of libzdb for index-rebuild
This method is really code-wise useful, it reuse all the existing system
about index and don't introduce new stuff, but this method rely on the
production index use, which in userkey mode can use a lot of memory if a
large amount of different keys are present.

In sequential mode, a lot of disk read/write are made, because of how
query works, but on SSD this is not an issue and rebuild can still be
quite fast, no high memory required.

This introduce a working history system, but rebuild is not yet perfect.
Only one datafile is proceed for now, and some few small bugs are still
present.