hook: clean zombies process
Hook subsystem was not waiting for child process to finish and previous
signal ignore set was disabled because of the wait introduced by missing
data event.

This lead to a lof of defunct process if hooks are triggered.

As soon as a new child is spawn, a counter is kept to known how many
expected child are still alive and a periodic call to
libzdb_hooks_cleanup will wait for them and discard them.