hook: check return code of 'ready' hook and abort on error
When using hook system, there is a 'ready' hook triggered. Previously,
whatever happened on this hook was ignored. Now, return code is checked
and any non-zero value will interrupt zdbd start.

Returning a non-zero value can be used to tell zdbd to stop
initialization because something external is not ready or safe to get
the database up and running.

This closes #100