hooks: rewrite hooks handling to support backlog list
Keep in memory a backlog hooks list.

This list will keep all current running hooks task and finished tasks
for certain amount of time (defined in hook.h in seconds).

This list can be queried to know which hooks is currently running and
which was executed and terminated. This list also keep return code,
status and arguments. It keep a timestamp when hook was created and when
it finished. The pid is included as well.

There is a fixed list allocated, this list grow if there is no empty
slot left. This list won't shrink.

Background runner will iterate over this list and check if some hooks
expired (timeout defined in hook.h), after that time, list is cleaned.