index: support dirty flags tracking
Index files are not immuable, they are the only mutable files on the
database to be able to keep delete/overwrite chain. For data files which
are immuable and really append-only, it's easy to do incremental backup,
but for index, some old files can have change by that time.

This change implement a way to keep track of which index files have
changed since last reset (or instance boot), a dirty map is kept in
memory and keep track of index changes.

This map use one bit per index files which make it quite small.