zdbd: force idle trigger after few commands
The background code executed when clients are idle (no socket event for
couple of ms) is only exected on idle. It's easy to send a simple
command to the server, every eg 100 ms. This will keep server busy and
never reach the timeout.

This would lead to never reach the background cleanup code.

This fix force redis_idle_process() to be called after 100 commands
received, whatever timeout was reached or not. This will ensure
background cleanup will always be triggered even on spamming.