zdbd: reopen stdin, stdout and stderr on /dev/null
There was a mistake in the previous workflow when closing stdin, stdout
and stderr when no external logfile were specified, making theses 3
files descriptors closed when they should not.

In place, they are redirected to /dev/null to always works as expected
but doing nothing. This is the expected behavior.

Closing 0, 1 and 2 were leading to data/socket corruption (obviously).