internal: change database files names (break compatibility)
Previous implementation was using old syntax dependant to limitation:
- index: namespace-name/zdb-index-XXXXX
- data: namespace-name/zdb-data-XXXXX

This was at the time when files was limited to 16 bits id and thus
limited to 65536 files. This limit is not anymore.

The new files names is way more compact:
- index: namespace-name/iX
- data: namespace-name/dX

There is still "i" and "d" prefix in order to know by filename if it's a
datafile or indexfile. The fileid is not zero-prefixed anymore.

This can reduce directory size dramatically by dropping lot of useless
redundent data (zdb-xxxx prefix is quite useless).