bootstrap: protect zdb instance with index/data lock
In order to avoid multiple instance of zdb using the same index or data
path (which would lead to corruption), there is a locking system
introduced.

When zdb starts, it put a lock on:
- DATADIR/.lockfile
- INDEXDIR/.lockfile

Theses locks are removed when zdb database is closed, stopped, crashed,
etc.

When trying to open a database with a acquired lock, open will be
denied.

In addition, namespace loader will only treat directories and not each
directory entries (it would use the lockfile as an entry otherwise).