libzdb: namespaces: new hook "namespace-closing" on emergency
Code block "emergency" (mainly used to flush descriptor in case of
crash, but it's always useful to flush even on gracefull stop) now
trigger a new hook called 'namespace-closing' with as arguments current
indexfile, current datafile and dirty index list.

This hook can be used to save current index/data and dirty list in order
to restore it later.

This implement #125