libzdb: remove creation/opened time on header
Keeping track of creation and last opened time in the data/index header
for each file was kinda interresting but not really useful (creation
date are still recorded on filesystem level).

But this had a major downside: restarting zdb were each time changing
the timestamp in the header, changing the checksum of the file, without
any real change to the data. This is annoying for incremental backup.

Header fields are kept as it, but timestamp are now forced to zero.
Restarting zdb don't change file checksum anymore.

This will be useful for coming upgrade with incremental replication.