core: delete: keep track of timestamp
Set current timestamp (or custom timestamp) when deleting a key.
Implementation accept a custom timestamp in libzdb.

In zdbd support, DEL command accept a new argument only valid by admin
user to set custom timestamp, if none, current timestamp is used.

This feature doesn't add any needs but keep track of when a key gets
deleted for history and tracking purpose. This data was kinda not needed
because never available (no way to retreive deleted data). This won't be
valid in further updates.