core: implement raw data fetcher
An internal call allows to retreive a raw data object from database only
based on fileid and offset. This method of access should only be made by
a valid fileid and offset, some protection are in place to avoid issues
on wrong offset but not fully tested yet.

This commit add a new command on zdbd: DATA RAW <fileid> <offset>

This command (only for admin) returns an array with the object
requested:
  1. Key
  2. Previous Offset
  3. Integrity CRC32
  4. Flags
  5. Timestamp
  6. Payload

In addition with latest NSINFO update, this command can be used to query
a database based on fielid/offset from another database used eg, for
replication.