libzdb: attempt to fix RSCAN on sequenrial mode
Issue #146 reports an issue using sequential mode and RSCAN not
returning sane values.

This is confirmed.

Code was still using old hardcoded fixed-length value (32 bits keys) not
relevant anymore, but moreover was still using a dirty fix needed
because of an old bug. This bug is probably not on the field anymore and
can be discarded, this commit bypass that old hotfix and use a more
classic/logic way how to deal RSCAN. Code will be cleaned after more
tests.